+++ backend-v3/configs/commonConfig.js
@@ -19,11 +19,11 @@ let baseConfig = {
   debugKeyHeader: 'x-mika-debug',
   debugKey: '1K24vDZGaGmJGCTTVIRyLxPPiHY',
 
-  allowedOrigin: isEnvProduction ? '*.mikaapp.id' : '*',
+  allowedOrigins: isEnvProduction ? [ 'https://backoffice.mikaapp.id' ] : '*',
 
   authSessionTokenHeader: 'x-access-token',
   authSecretKey: '4FoC5uMLqAkoBMSw2sOLIF7M',
-  authExpirySecond: 15 * 60 * 24,
+  authExpirySecond: 4 * 24 * 3600,
 
   transactionExpirySecond: 3 * 60,